project has latest changes whereby the image is not being done as a microservice
