--- image_service/src/main/java/com/foodapi/image_service/ImageServiceApplication.java
@@ -1,15 +1,0 @@
-package com.foodapi.image_service;
-
-import com.foodapi.image_service.service.ImageService;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-
-@SpringBootApplication
-public class ImageServiceApplication {
-
-    public static void main(String[] args)
-    {
-        SpringApplication.run(ImageServiceApplication.class,args);
-    }
-}
