--- image_service/src/main/java/com/foodapi/image_service/repo/ImageRepo.java
@@ -1,9 +1,0 @@
-package com.foodapi.image_service.repo;
-
-import com.foodapi.image_service.model.ImageModel;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ImageRepo extends JpaRepository<ImageModel,Long> {
-}
