--- image_service/src/main/java/com/foodapi/image_service/service/ImageService.java
@@ -1,4 +1,0 @@
-package com.foodapi.image_service.service;
-
-public interface ImageService {
-}
