--- image_service/src/main/resources/application.properties
@@ -1,16 +1,0 @@
-server.port=8080
-logging.file.path=/var/tmp/mylog.log
-spring.datasource.url=jdbc:postgresql://localhost:5433/food_db
-spring.datasource.username=admin
-spring.datasource.password=12345
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.hibernate.show-sql=true
-spring.datasource.tomcat.test-while-idle=true
-spring.datasource.tomcat.validation-query=SELECT 1
-spring.sql.init.mode=always
-
-cloud.name=dywf1gnyt
-cloud.api.key=477736346853843
-cloud.api.secret=4ycwXnGpQe_e6aKc82ddqcMROXY
